requirements.txt
flask
flask-cors
pillow
gunicorn
